add grid (base class)
- `structured_grid` inherits from `grid`: it only adds
  a few properties (spacing, length, etc.)
- Move most of structured_grid.hpp into grid.hpp
