+++ test/test_raster_grid.cpp
